Bug fixes for User Ranking entities. Logging for Reporting Job. Fix for precision to 2 decimal places for rating.

+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/entities/UserRankingPastMonthRegion.java
@@ -72,7 +72,7 @@ public class UserRankingPastMonthRegion
     private int totalReviews;
     
     @Column(name = "average_rating")
-    private int averageRating;
+    private float averageRating;
     
     @Column(name = "is_eligible")
     private	int isEligible;
@@ -86,12 +86,12 @@ public class UserRankingPastMonthRegion
     @Transient
     private String profileImageUrlThumbnail;
     
-	public int getAverageRating() {
+	
+	public float getAverageRating() {
 		return averageRating;
 	}
 
-
-	public void setAverageRating(int averageRating) {
+	public void setAverageRating(float averageRating) {
 		this.averageRating = averageRating;
 	}
